+++ src/src/WebServer/LoadFromFS.cpp
@@ -160,9 +160,9 @@ size_t streamFromFS(String path, bool htmlEscape) {
       chunksize = available;
     }
     uint8_t buf[64] = {0};
-    const size_t read = f.read(buf, chunksize);
+    const int read = f.read(buf, chunksize);
     if (read == chunksize) {
-      for (uint32_t i = 0; i < chunksize; ++i) {
+      for (int32_t i = 0; i < chunksize; ++i) {
         const char c = (char)buf[i];
         if (htmlEscape && htmlEscapeChar(c, escaped)) {
           addHtml(escaped);
